float intensity all, custom apo still needs fixing

+++ ConsoleApplication/src/ImageFacilities.h
@@ -35,10 +35,10 @@ bool importImageCV(std::vector<tracer::PI_LuminousPoint>& outputvec, std::string
 
 //bool importImageCVv2(std::vector<tracer::PI_LuminousPoint>& outputvec, std::string path, float posX, float posY, float posZ, float sizeHorz, float sizeVert, float rotX, float rotY, float rotZ, float brightness);
 
-bool importCustomApo(double*& p_customApoData, int& customApoDataSize, std::string path);
+bool importCustomApo(float*& p_customApoData, int& customApoDataSize, std::string path);
 
 
-bool generateGLDrawTexture(unsigned char*& output, double* input, int& rows, int& cols);
+bool generateGLDrawTexture(unsigned char*& output, float* input, int& rows, int& cols);
 //this function will maintain its own texture cache, after use, please call the clear function below
 void clearGLDrawTexture();
